% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binProb.R
\name{binProb}
\alias{binProb}
\title{Binomial CDF calculator}
\usage{
binProb(n, p, min, max)
}
\arguments{
\item{n}{number of bernoulli trials}

\item{p}{probability p for each bernoulli trial}

\item{min}{minimum number of successes}

\item{max}{maximum number of successes}
}
\value{
A probability
}
\description{
Calculates the probability that min <= X <= max, given X~Bin(n,p).
}
\examples{
binProb(7, 0.3, 1, 3)
}
